Cambio de int a float para tener el costo total del viaje.

+++ src/Moto.java
@@ -2,23 +2,22 @@ import java.util.Scanner;
 
 public class Moto {
 
-    public int viajeMoto() {
+    public float viajeMoto() {
         Scanner sc = new Scanner(System.in);
-
         System.out.println("Ingresa la ciudad de destino:");
         String ciudadDestino = sc.nextLine();
 
         System.out.println("Ingresa la distancia recorrida en km:");
-        int distanciaKm = sc.nextInt();
+        float distanciaKm = sc.nextInt();
 
         System.out.println("Ingresa el peaje pagado (0 si no hubo):");
-        int peaje = sc.nextInt();
+        float peaje = sc.nextInt();
 
         System.out.println("Ingresa el costo de la gasolina por galón:");
-        int costoGasolina = sc.nextInt();
+        float costoGasolina = sc.nextInt();
 
-        int consumoGalon = distanciaKm / 100;  // 100 km por galón (ejemplo)
-        int gastoGasolina = consumoGalon * costoGasolina;
+        float consumoGalon = distanciaKm / 100;  // 100 km por galón (ejemplo)
+        float gastoGasolina = consumoGalon * costoGasolina;
 
         System.out.println("-------------");
         System.out.println("Destino: " + ciudadDestino);
